+++ src/app/api/auth/callback/route.ts
@@ -1,0 +1,10 @@
+import { NextResponse } from 'next/server';
+
+export const GET = async (request: Request) => {
+  const { searchParams, origin } = new URL(request.url);
+  const error = searchParams.get('error');
+  if (error) {
+    return NextResponse.redirect(`${origin}/login`);
+  }
+  return NextResponse.redirect(`${origin}/signup`);
+};
