+++ src/app/signup/page.tsx
@@ -1,0 +1,5 @@
+const SignupPage = () => {
+  return <div>SignupPage</div>;
+};
+
+export default SignupPage;
